Move test to data tests
